@@ -36,27 +30,29 @@
         <module>alphatimer-server</module>
         <module>alphatimer-livetiming</module>
 
-        <module>alphatimer-package</module>
+        <module>alphatimer-distribution</module>
     </modules>
 
     <name>AlphaTimer</name>
-    <url>https://github.com/dennisfabri/alpha-translator</url>
+    <url>https://github.com/LISASP-org/alphatimer</url>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>11</java.version>
-        <maven.compiler.release>11</maven.compiler.release>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <java.version>17</java.version>
+        <maven.compiler.release>${java.version}</maven.compiler.release>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
+
+        <basics.version>1.0.0-SNAPSHOT</basics.version>
 
         <slf4j.version>1.7.32</slf4j.version>
-        <kotlin.version>1.5.30</kotlin.version>
         <lombok.version>1.18.20</lombok.version>
         <nrjavaserial.version>5.2.1</nrjavaserial.version>
         <xstream.version>1.4.18</xstream.version>
-        <vaadin.version>20.0.7</vaadin.version>
-        <vaadin.helper.version>1.7.1</vaadin.helper.version>
+        <vaadin.version>21.0.0</vaadin.version>
+        <vaadin.helper.version>1.7.2</vaadin.helper.version>
         <vaadin.exampledata.version>3.4.0</vaadin.exampledata.version>
+        <jetbrains.annotations.version>16.0.2</jetbrains.annotations.version>
 
         <junit-jupiter.version>5.7.2</junit-jupiter.version>
         <junit.version>4.13.2</junit.version>
