@@ -1,134 +1,82 @@
 package org.lisasp.alphatimer.test.datatests;
 
-import org.junit.jupiter.api.*;
+import lombok.Value;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.lisasp.alphatimer.api.refinedmessages.RefinedMessageListener;
 import org.lisasp.alphatimer.api.refinedmessages.accepted.*;
 import org.lisasp.alphatimer.api.refinedmessages.dropped.*;
 import org.lisasp.alphatimer.datatests.TestData;
-import org.lisasp.alphatimer.jre.date.DateTimeFacade;
+import org.lisasp.basics.jre.date.DateTimeFacade;
 import org.lisasp.alphatimer.protocol.InputCollector;
 import org.lisasp.alphatimer.protocol.MessageAggregator;
 import org.lisasp.alphatimer.protocol.MessageConverter;
 import org.lisasp.alphatimer.refinedmessages.DataHandlingMessageRefiner;
 import org.lisasp.alphatimer.test.datatests.testdoubles.DateTimeFacadeTestDouble;
-import org.mockito.Mockito;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
-import java.util.Hashtable;
 
-import static org.mockito.Mockito.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class RefinedMessagesDataTest {
-
-    private static Hashtable<Class<?>, Integer> createHashtable(int didNotFinish,
-                                                                   int didNotStart,
-                                                                   int finish,
-                                                                   int officialEnd,
-                                                                   int readyToStart,
-                                                                   int start,
-                                                                   int takeOverTime,
-                                                                   int usedLanes) {
-        Hashtable<Class<?>, Integer> hashtable = new Hashtable<>();
-
-        hashtable.put(DidNotFinishMessage.class, didNotFinish);
-        hashtable.put(DidNotStartMessage.class, didNotStart);
-        hashtable.put(TimeMessage.class, finish);
-        hashtable.put(OfficialEndMessage.class, officialEnd);
-        hashtable.put(ReadyToStartMessage.class, readyToStart);
-        hashtable.put(StartMessage.class, start);
-        hashtable.put(TakeOverTimeMessage.class, takeOverTime);
-        hashtable.put(UsedLanesMessage.class, usedLanes);
-
-        return hashtable;
-    }
-
-    private static void extendHashtable(Hashtable<Class<?>, Integer> hashtable,
-                                        int didNotFinish,
-                                        int didNotStart,
-                                        int finish,
-                                        int officialEnd,
-                                        int readyToStart,
-                                        int start,
-                                        int takeOverTime,
-                                        int usedLanes,
-                                        int droppedUnknown) {
-
-        hashtable.put(DroppedDidNotFinishMessage.class, didNotFinish);
-        hashtable.put(DroppedDidNotStartMessage.class, didNotStart);
-        hashtable.put(DroppedTimeMessage.class, finish);
-        hashtable.put(DroppedOfficialEndMessage.class, officialEnd);
-        hashtable.put(DroppedReadyToStartMessage.class, readyToStart);
-        hashtable.put(DroppedStartMessage.class, start);
-        hashtable.put(DroppedTakeOverTimeMessage.class, takeOverTime);
-        // hashtable.put(DroppedUsedLanesMessage.class, usedLanes);
-
-        hashtable.put(DroppedUnknownMessage.class, droppedUnknown);
-    }
-
-    private static final Hashtable<String, Hashtable<Class<?>, Integer>> testData = new Hashtable<>();
-
-    private static final Hashtable<Class<?>, Integer> EMPTY = new Hashtable<>();
-
-    private static int getValue(String name, Class<?> message) {
-        return testData.getOrDefault(name, EMPTY).getOrDefault(message, 0);
-    }
-
-    static {
-        testData.put("DM2008Freitag", createHashtable(0, 0, 899, 0, 67, 67, 373, 1406));
-        extendHashtable(testData.get("DM2008Freitag"), 0, 0, 0, 0, 0, 0, 0, 0, 0);
-        testData.put("DM2008Samstag", createHashtable(0, 0, 1177, 0, 154, 145, 0, 1476));
-        extendHashtable(testData.get("DM2008Samstag"), 0, 0, 0, 0, 0, 0, 0, 0, 0);
-
-        testData.put("DM2009", createHashtable(0, 0, 5739, 8, 658, 638, 1254, 8011));
-        extendHashtable(testData.get("DM2009"), 0, 0, 2, 0, 0, 2, 0, 0, 0);
-
-        testData.put("DM2010", createHashtable(0, 0, 1621, 2, 147, 142, 525, 2435));
-        extendHashtable(testData.get("DM2010"), 0, 0, 0, 0, 0, 0, 0, 0, 0);
-
-        testData.put("JRP2019Freitag", createHashtable(0, 0, 0, 0, 2, 1, 0, 3));
-        extendHashtable(testData.get("JRP2019Freitag"), 0, 0, 0, 0, 0, 0, 0, 0, 0);
-        testData.put("JRP2019Samstag", createHashtable(0, 47, 846, 97, 109, 100, 0, 1055));
-        extendHashtable(testData.get("JRP2019Samstag"), 0, 0, 0, 0, 0, 0, 0, 0, 754);
-        testData.put("JRP2019Sonntag", createHashtable(2, 29, 533, 51, 69, 57, 0, 659));
-        extendHashtable(testData.get("JRP2019Sonntag"), 0, 0, 0, 0, 0, 0, 0, 0, 408);
-
-        testData.put("DP2019Freitag", createHashtable(0, 0, 1226, 86, 96, 94, 0, 1416));
-        extendHashtable(testData.get("DP2019Freitag"), 0, 0, 0, 0, 0, 0, 0, 0, 2);
-        testData.put("DP2019Samstag", createHashtable(0, 0, 1220, 68, 75, 73, 0, 1368));
-        extendHashtable(testData.get("DP2019Samstag"), 0, 0, 0, 0, 0, 0, 0, 0, 0);
-
-        testData.put("DMM2019Freitag", createHashtable(0, 0, 9, 0, 3, 2, 0, 14));
-        extendHashtable(testData.get("DMM2019Freitag"), 0, 0, 0, 0, 0, 0, 0, 0, 0);
-        testData.put("DMM2019Samstag", createHashtable(0, 0, 1535, 120, 139, 134, 0, 1811));
-        extendHashtable(testData.get("DMM2019Samstag"), 0, 0, 3, 0, 0, 0, 0, 0, 5);
-        testData.put("DMM2019Sonntag", createHashtable(0, 0, 1219, 140, 163, 164, 0, 1546));
-        extendHashtable(testData.get("DMM2019Sonntag"), 0, 0, 0, 0, 0, 0, 0, 0, 0);
-    }
 
     private static final boolean verbose = false;
 
     private InputCollector inputCollector;
-    private RefinedMessageListener listener;
-    private DataHandlingMessageRefiner refiner;
+    private HitCounter counter;
 
-    @BeforeAll
-    static void prepareData() throws IOException {
-        new TestData().prepare();
+    @Value
+    static class TestDescription {
+        private final String name;
+        private final int didNotFinish;
+        private final int didNotStart;
+        private final int time;
+        private final int officialEnd;
+        private final int readyToStart;
+        private final int start;
+        private final int takeOverTime;
+        private final int usedLanes;
+        private final int droppedDidNotFinish;
+        private final int droppedDidNotStart;
+        private final int droppedTime;
+        private final int droppedOfficialEnd;
+        private final int droppedReadyToStart;
+        private final int droppedStart;
+        private final int droppedTakeOverTime;
+        private final int droppedUsedLanes;
+        private final int droppedUnknown;
+
+        @Override
+        public String toString() {
+            return name;
+        }
     }
 
-    @AfterAll
-    static void cleanup() throws IOException {
-        new TestData().cleanup();
+    static TestDescription[] provideTestDescriptions() {
+        return new TestDescription[]{
+                new TestDescription("DM2008-Freitag", 0, 0, 899, 0, 67, 67, 373, 1406, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+                new TestDescription("DM2008-Samstag", 0, 0, 1177, 0, 154, 145, 0, 1476, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+                new TestDescription("DM2009", 0, 0, 5739, 8, 658, 638, 1254, 8011, 0, 0, 2, 0, 0, 2, 0, 0, 0),
+                new TestDescription("DM2010", 0, 0, 1621, 2, 147, 142, 525, 2435, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+                new TestDescription("DMM2019-Freitag", 0, 0, 9, 0, 3, 2, 0, 14, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+                new TestDescription("DMM2019-Samstag", 0, 0, 1535, 120, 139, 134, 0, 1811, 0, 0, 3, 0, 0, 0, 0, 0, 5),
+                new TestDescription("DMM2019-Sonntag", 0, 0, 1219, 140, 163, 164, 0, 1546, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+                new TestDescription("DP2019-Freitag", 0, 0, 1226, 86, 96, 94, 0, 1416, 0, 0, 0, 0, 0, 0, 0, 0, 2),
+                new TestDescription("DP2019-Samstag", 0, 0, 1220, 68, 75, 73, 0, 1368, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+                new TestDescription("DEM2021-Samstag", 0, 0, 414, 23, 30, 29, 0, 576, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+                new TestDescription("DEM2021-Sonntag", 0, 0, 764, 45, 67, 57, 0, 1054, 0, 0, 0, 0, 0, 0, 0, 0, 5),
+        };
     }
 
     @BeforeEach
     void prepare() {
-        listener = mock(RefinedMessageListener.class);
+        counter = new HitCounter();
+        RefinedMessageListener listener = refinedMessage -> counter.increaseHitCount(refinedMessage.getClass());
 
-        refiner = new DataHandlingMessageRefiner();
+        DataHandlingMessageRefiner refiner = new DataHandlingMessageRefiner();
         refiner.register(listener);
 
         DateTimeFacade dateTimeFacade = new DateTimeFacadeTestDouble(LocalDateTime.of(2021, 6, 1, 10, 0));
@@ -153,46 +101,37 @@ class RefinedMessagesDataTest {
     @AfterEach
     void cleanUp() {
         inputCollector = null;
-        listener = null;
     }
 
     @ParameterizedTest
-    @ValueSource(strings = {
-            "DM2008Freitag", "DM2008Samstag",
-            "DM2009",
-            "DM2010",
-            "JRP2019Freitag", "JRP2019Samstag", "JRP2019Sonntag",
-            "DP2019Freitag", "DP2019Samstag",
-            "DMM2019Freitag", "DMM2019Samstag", "DMM2019Sonntag"})
-    void test(String name) throws IOException {
-        byte[] data = new TestData().readSerialInput(name);
+    @MethodSource("provideTestDescriptions")
+    void test(TestDescription description) throws IOException {
+        byte[] data = new TestData().readSerialInput(description.getName());
 
         for (byte d : data) {
             inputCollector.accept(d);
         }
         inputCollector.close();
 
+        assertEquals(description.getDidNotFinish(), counter.getHitCount(DidNotFinishMessage.class), "DidNotFinishMessage");
+        assertEquals(description.getDidNotStart(), counter.getHitCount(DidNotStartMessage.class), "DidNotStartMessage");
+        assertEquals(description.getTime(), counter.getHitCount(TimeMessage.class), "TimeMessage");
+        assertEquals(description.getOfficialEnd(), counter.getHitCount(OfficialEndMessage.class), "OfficialEndMessage");
+        assertEquals(description.getReadyToStart(), counter.getHitCount(ReadyToStartMessage.class), "ReadyToStartMessage");
+        assertEquals(description.getStart(), counter.getHitCount(StartMessage.class), "StartMessage");
+        assertEquals(description.getTakeOverTime(), counter.getHitCount(TakeOverTimeMessage.class), "TakeOverTimeMessage");
+        assertEquals(description.getUsedLanes(), counter.getHitCount(UsedLanesMessage.class), "UsedLanesMessage");
 
-        verify(listener, times(getValue(name, DidNotFinishMessage.class))).accept(Mockito.any(DidNotFinishMessage.class));
-        verify(listener, times(getValue(name, DidNotStartMessage.class))).accept(Mockito.any(DidNotStartMessage.class));
-        verify(listener, times(getValue(name, TimeMessage.class))).accept(Mockito.any(TimeMessage.class));
-        verify(listener, times(getValue(name, OfficialEndMessage.class))).accept(Mockito.any(OfficialEndMessage.class));
-        verify(listener, times(getValue(name, ReadyToStartMessage.class))).accept(Mockito.any(ReadyToStartMessage.class));
-        verify(listener, times(getValue(name, StartMessage.class))).accept(Mockito.any(StartMessage.class));
-        verify(listener, times(getValue(name, TakeOverTimeMessage.class))).accept(Mockito.any(TakeOverTimeMessage.class));
-        verify(listener, times(getValue(name, UsedLanesMessage.class))).accept(Mockito.any(UsedLanesMessage.class));
+        assertEquals(description.getDroppedDidNotFinish(), counter.getHitCount(DroppedDidNotFinishMessage.class), "DroppedDidNotFinishMessage");
+        assertEquals(description.getDroppedDidNotStart(), counter.getHitCount(DroppedDidNotStartMessage.class), "DroppedDidNotStartMessage");
+        assertEquals(description.getDroppedTime(), counter.getHitCount(DroppedTimeMessage.class), "DroppedTimeMessage");
+        assertEquals(description.getDroppedOfficialEnd(), counter.getHitCount(DroppedOfficialEndMessage.class), "DroppedOfficialEndMessage");
+        assertEquals(description.getDroppedReadyToStart(), counter.getHitCount(DroppedReadyToStartMessage.class), "DroppedReadyToStartMessage");
+        assertEquals(description.getDroppedStart(), counter.getHitCount(DroppedStartMessage.class), "DroppedStartMessage");
+        assertEquals(description.getDroppedTakeOverTime(), counter.getHitCount(DroppedTakeOverTimeMessage.class), "DroppedTakeOverTimeMessage");
 
-        verify(listener, times(getValue(name, DroppedDidNotFinishMessage.class))).accept(Mockito.any(DroppedDidNotFinishMessage.class));
-        verify(listener, times(getValue(name, DroppedDidNotStartMessage.class))).accept(Mockito.any(DroppedDidNotStartMessage.class));
-        verify(listener, times(getValue(name, DroppedTimeMessage.class))).accept(Mockito.any(DroppedTimeMessage.class));
-        verify(listener, times(getValue(name, DroppedOfficialEndMessage.class))).accept(Mockito.any(DroppedOfficialEndMessage.class));
-        verify(listener, times(getValue(name, DroppedReadyToStartMessage.class))).accept(Mockito.any(DroppedReadyToStartMessage.class));
-        verify(listener, times(getValue(name, DroppedStartMessage.class))).accept(Mockito.any(DroppedStartMessage.class));
-        verify(listener, times(getValue(name, DroppedTakeOverTimeMessage.class))).accept(Mockito.any(DroppedTakeOverTimeMessage.class));
+        assertEquals(description.getDroppedUnknown(), counter.getHitCount(DroppedUnknownMessage.class), "DroppedUnknownMessage");
 
-        verify(listener, times(getValue(name, DroppedUnknownMessage.class))).accept(Mockito.any(DroppedUnknownMessage.class));
-
-        verifyNoMoreInteractions(listener);
+        counter.noMoreHits();
     }
-
 }
