@@ -9,9 +9,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.lisasp.alphatimer.api.protocol.json.BytesInputEventModule;
 import org.lisasp.alphatimer.heats.service.DataRepository;
 import org.lisasp.alphatimer.heats.service.HeatService;
-import org.lisasp.alphatimer.jre.date.ActualDateTime;
-import org.lisasp.alphatimer.jre.date.DateTimeFacade;
-import org.lisasp.alphatimer.spring.jms.JsonMessageConverter;
+import org.lisasp.basics.jre.date.ActualDateTime;
+import org.lisasp.basics.jre.date.DateTimeFacade;
+import org.lisasp.basics.spring.jms.JsonMessageConverter;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
