@@ -1,10 +1,13 @@
 package org.lisasp.alphatimer.messagesstorage;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.lisasp.alphatimer.api.protocol.events.messages.enums.*;
-import org.lisasp.alphatimer.spring.jpa.VersionedBaseEntity;
+import org.lisasp.basics.spring.jpa.BaseEntity;
 
 import javax.persistence.*;
+import javax.validation.constraints.Min;
 import java.time.LocalDateTime;
 
 @Entity()
@@ -12,7 +15,7 @@ import java.time.LocalDateTime;
 @Table(name = "aresmessage")
 @NoArgsConstructor
 @AllArgsConstructor
-public class AresMessage extends VersionedBaseEntity {
+public class AresMessage extends BaseEntity {
     @Column(nullable = false)
     private String originalText1;
     @Column(nullable = false)
@@ -30,15 +33,22 @@ public class AresMessage extends VersionedBaseEntity {
     private TimeType timeType;
     @Column(nullable = false)
     private String usedLanes;
+    @Min(0)
     private byte lapCount;
+    @Min(0)
     private short event;
+    @Min(0)
     private byte heat;
+    @Min(0)
     private byte rank;
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
     private RankInfo rankInfo;
+    @Min(0)
     private byte lane;
+    @Min(0)
     private byte currentLap;
+    @Min(0)
     private int timeInMillis;
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
