@@ -1,9 +1,10 @@
 package org.lisasp.alphatimer.messagesstorage;
 
-import org.springframework.data.repository.CrudRepository;
-
 import java.util.List;
 
-public interface AresMessageRepository extends CrudRepository<AresMessage, String> {
+public interface AresMessageRepository {
     List<AresMessage> findAllByCompetitionKeyAndEventAndHeat(String competitionKey, short event, byte heat);
+
+    <S extends AresMessage> S save(S s);
+    long count();
 }
