@@ -8,7 +8,7 @@ import org.lisasp.alphatimer.api.protocol.events.dropped.UnstructuredInputDroppe
 import org.lisasp.alphatimer.api.protocol.events.messages.DataHandlingMessage2;
 import org.lisasp.alphatimer.api.protocol.events.messages.enums.TimeInfo;
 import org.lisasp.alphatimer.api.protocol.events.messages.enums.TimeMarker;
-import org.lisasp.alphatimer.jre.date.DateTimeFacade;
+import org.lisasp.basics.jre.date.DateTimeFacade;
 import org.lisasp.alphatimer.protocol.InputCollector;
 import org.lisasp.alphatimer.protocol.MessageConverter;
 import org.mockito.Mockito;
