@@ -7,7 +7,7 @@ import org.lisasp.alphatimer.api.protocol.Characters;
 import org.lisasp.alphatimer.api.protocol.DataInputEventListener;
 import org.lisasp.alphatimer.api.protocol.events.dropped.UnknownMessageDroppedEvent;
 import org.lisasp.alphatimer.api.protocol.events.messages.Ping;
-import org.lisasp.alphatimer.jre.date.DateTimeFacade;
+import org.lisasp.basics.jre.date.DateTimeFacade;
 import org.lisasp.alphatimer.protocol.InputCollector;
 import org.lisasp.alphatimer.protocol.MessageConverter;
 import org.mockito.Mockito;
