@@ -4,9 +4,7 @@ module org.lisasp.alphatimer.refinedmessages {
     requires transitive org.lisasp.alphatimer.api.refinedmessages;
     requires transitive org.lisasp.alphatimer.api.protocol;
 
-    requires org.lisasp.alphatimer.messaging;
-
-    requires java.base;
+    requires org.lisasp.basics.notification;
 
     requires static lombok;
 }
