@@ -1,6 +1,6 @@
 package org.lisasp.alphatimer.serial;
 
-import org.lisasp.alphatimer.messaging.ByteListener;
+import org.lisasp.basics.notification.primitive.ByteListener;
 
 public interface SerialPortReader extends AutoCloseable {
     SerialPortReader register(ByteListener listener);
