@@ -2,8 +2,8 @@ package org.lisasp.alphatimer.serial;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import org.lisasp.alphatimer.jre.date.DateFacade;
-import org.lisasp.alphatimer.jre.io.FileFacade;
+import org.lisasp.basics.jre.date.DateFacade;
+import org.lisasp.basics.jre.io.FileFacade;
 
 import java.io.IOException;
 import java.time.format.DateTimeFormatter;
