@@ -3,16 +3,16 @@ package org.lisasp.alphatimer.serialportlistener;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.lisasp.alphatimer.api.protocol.json.BytesInputEventModule;
-import org.lisasp.alphatimer.jre.date.ActualDate;
-import org.lisasp.alphatimer.jre.date.ActualDateTime;
-import org.lisasp.alphatimer.jre.date.DateFacade;
-import org.lisasp.alphatimer.jre.date.DateTimeFacade;
-import org.lisasp.alphatimer.jre.io.ActualFile;
+import org.lisasp.basics.jre.date.ActualDate;
+import org.lisasp.basics.jre.date.ActualDateTime;
+import org.lisasp.basics.jre.date.DateFacade;
+import org.lisasp.basics.jre.date.DateTimeFacade;
+import org.lisasp.basics.jre.io.ActualFile;
 import org.lisasp.alphatimer.protocol.InputCollector;
 import org.lisasp.alphatimer.serial.DefaultSerialConnectionBuilder;
 import org.lisasp.alphatimer.serial.SerialConnectionBuilder;
 import org.lisasp.alphatimer.serial.Storage;
-import org.lisasp.alphatimer.spring.jms.JsonMessageConverter;
+import org.lisasp.basics.spring.jms.JsonMessageConverter;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
