@@ -1,6 +1,6 @@
 package org.lisasp.alphatimer.server.testdoubles;
 
-import org.lisasp.alphatimer.jre.date.DateFacade;
+import org.lisasp.basics.jre.date.DateFacade;
 
 import java.time.LocalDate;
 import java.time.Month;
